requirements.txt
Flask
google-api-python-client
google-auth
google-auth-oauthlib
google-cloud-storage
google-cloud-secret-manager
gunicorn
